requirements.txt
yaml
Django==2.0.9
django-rest-framework-mongoengine==3.3.1    
djangorestframework==3.9.0    
djangorestframework-jsonapi==2.6.0    
djangorestframework-jwt==1.11.0   
djangotoolbox==1.8.0
django-rest-swagger==2.2.0
mongoengine==0.9.0
pymongo==2.8
dockerfile-parse==0.0.11
ruamel.yaml==0.15.77
ruamel.yaml.jinja2==0.2.2
PyYAML==3.13
